New hierarchy and method
'Instance' was replaced with 'Container_Instance,' which is a combined ROBLOX file of the main module and the configuration children.

New module called 'source,' which is the main code of the module. This will be easier to replace per update, and will exist as an immediate child of the main module container